Squashed commit of the following:
commit 46ee1b7c3f8d8741863151eeec45071b252bc3cd
Author: Markus Dreseler <markus.dreseler@hpi.de>
Date:   Mon Apr 10 10:33:26 2017 +0200

    Merged Nils' gs 101-rewrite-sort-to-be-always-materializing

+++ src/lib/storage/index/adaptive_radix_tree/adaptive_radix_tree_nodes.cpp
@@ -374,7 +374,7 @@ BaseIndex::Iterator Node256::begin() const {
 }
 
 BaseIndex::Iterator Node256::end() const {
-  for (uint8_t i = _children.size() - 1; i >= 0; --i) {
+  for (int16_t i = _children.size() - 1; i >= 0; --i) {
     if (_children[i] != nullptr) {
       return _children[i]->begin();
     }
